requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tensorflow-cpu>=2.15.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
pytest>=7.4.0
